fix: improve selector logic
Commit change containers considered to marked when focus change event
happens and resets mark on closed containers.

In the "focus" case, we require type == "con" and having a pid
associated with it.

In the "close" case, we clear marks if they match the event's container
id.

Changes reduce chance of trying to update marks on already removed
containers and only adding marks on containers.